fix bug in parsing json.

+++ external/raw_transaction.go
@@ -3,31 +3,31 @@ package external
 import "tax-management/taxDep/types"
 
 type AfterData struct {
-	Taxid   string `json:"taxid"`
-	Tinb    string `json:"tinb"`
-	Tins    string `json:"tins"`
-	Adis    int64  `json:"adis"`
-	Am      int64  `json:"am"`
-	Dis     int64  `json:"dis"`
-	Fee     int64  `json:"fee"`
-	Indatim int64  `json:"indatim"`
-	Inp     int64  `json:"inp"`
-	Ins     int64  `json:"ins"`
-	Inty    int64  `json:"inty"`
-	Prdis   int64  `json:"prdis"`
-	Setm    int64  `json:"setm"`
-	Sstid   string `json:"sstid"`
-	Tadis   int64  `json:"tadis"`
-	Tbill   int64  `json:"tbill"`
-	Tdis    int64  `json:"tdis"`
-	Tob     int64  `json:"tob"`
-	Todam   int64  `json:"todam"`
-	Tprdis  int64  `json:"tprdis"`
-	Trn     string `json:"trn"`
-	Tsstam  int64  `json:"tsstam"`
-	Tvam    int64  `json:"tvam"`
-	Vam     int64  `json:"vam"`
-	Vra     int64  `json:"vra"`
+	Taxid   string  `json:"taxid"`
+	Tinb    string  `json:"tinb"`
+	Tins    string  `json:"tins"`
+	Adis    float64 `json:"adis"`
+	Am      float64 `json:"am"`
+	Dis     float64 `json:"dis"`
+	Fee     float64 `json:"fee"`
+	Indatim int64   `json:"indatim"`
+	Inp     int64   `json:"inp"`
+	Ins     int64   `json:"ins"`
+	Inty    int64   `json:"inty"`
+	Prdis   float64 `json:"prdis"`
+	Setm    int64   `json:"setm"`
+	Sstid   string  `json:"sstid"`
+	Tadis   float64 `json:"tadis"`
+	Tbill   float64 `json:"tbill"`
+	Tdis    float64 `json:"tdis"`
+	Tob     int64   `json:"tob"`
+	Todam   float64 `json:"todam"`
+	Tprdis  float64 `json:"tprdis"`
+	Trn     string  `json:"trn"`
+	Tsstam  float64 `json:"tsstam"`
+	Tvam    float64 `json:"tvam"`
+	Vam     float64 `json:"vam"`
+	Vra     float64 `json:"vra"`
 }
 type SourceData struct {
 	Connector string      `json:"connector"`
